refactor: Legacy MVC 제거하기
