--- app/src/main/java/com/techcourse/ManualHandlerMapping.java
@@ -1,34 +1,0 @@
-package com.techcourse;
-
-import jakarta.servlet.http.HttpServletRequest;
-import java.util.HashMap;
-import java.util.Map;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-import webmvc.org.springframework.web.servlet.HandlerMapping;
-import webmvc.org.springframework.web.servlet.mvc.asis.Controller;
-import webmvc.org.springframework.web.servlet.mvc.asis.ForwardController;
-
-public class ManualHandlerMapping implements HandlerMapping {
-
-    private static final Logger log = LoggerFactory.getLogger(ManualHandlerMapping.class);
-
-    private static final Map<String, Controller> controllers = new HashMap<>();
-
-    @Override
-    public void initialize() {
-        controllers.put("/", new ForwardController("/index.jsp"));
-
-        log.info("Initialized Handler Mapping!");
-        controllers.keySet()
-                .forEach(path -> log.info("Path : {}, Controller : {}", path, controllers.get(path).getClass()));
-    }
-
-    @Override
-    public Controller getHandler(HttpServletRequest request) {
-        String requestURI = request.getRequestURI();
-        log.debug("Request Mapping Uri : {}", requestURI);
-
-        return controllers.get(requestURI);
-    }
-}
